Add files via upload
Adaptation of Dijkstra's shortest path using lists that allows for the removal of 1 wall that will result in the shortest possible path from array[0][0] to array[length-1][length-1]. Note this was used on paths with a guarantee path from one end to the other. Concept idea from foobar with google.